Add files via upload
r are trainer nodes; e are predictor nodes; m is the manager node; boot is the bootstrap file to run the whole network; deep learning files are Jupyter files that run deep learning codes. Each file is associated with a blockchain node address and has access to the IPFS network.